requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioresponses==0.7.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncio==3.4.3
atomicwrites==1.4.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
distro==1.9.0
docutils==0.21.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
jiter==0.6.1
keyring==25.5.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.5.0
multidict==6.1.0
nh3==0.2.18
openai==1.52.2
packaging==24.1
pkginfo==1.10.0
pluggy==1.5.0
propcache==0.2.0
py==1.11.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-runner==6.0.1
python-dotenv==1.0.1
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.3
SecretStorage==3.3.3
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.6
twine==5.1.1
typing_extensions==4.12.2
urllib3==2.2.3
wheel==0.44.0
yarl==1.16.0
zipp==3.20.2
